third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-objects-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
Tests MediaSource, SourceBuffer, and SourceBufferList objects.

EVENT(webkitsourceopen)
Test MediaSource object type
EXPECTED (mediaSource == '[object WebKitMediaSource]') OK
EXPECTED (mediaSource instanceof window.WebKitMediaSource == 'true') OK

Add a SourceBuffer
EXPECTED (mediaSource.sourceBuffers.length == '0') OK
RUN(segmentHelper.addSourceBuffer())
EXPECTED (mediaSource.sourceBuffers.length == '1') OK

Test SourceBufferList object type:
EXPECTED (mediaSource.sourceBuffers == '[object WebKitSourceBufferList]') OK
EXPECTED (mediaSource.sourceBuffers instanceof window.WebKitSourceBufferList == 'true') OK

Test SourceBuffer object type
EXPECTED (segmentHelper.sourceBuffer instanceof window.WebKitSourceBuffer == 'true') OK
EXPECTED (mediaSource.sourceBuffers[0] instanceof window.WebKitSourceBuffer == 'true') OK

Append init segment
EXPECTED (mediaSource.sourceBuffers[0].buffered.length == '0') OK

Append 1 second of data.
EXPECTED (mediaSource.sourceBuffers[0].buffered.length == '1') OK
EXPECTED (mediaSource.sourceBuffers[0].buffered.start(0) == '0') OK
EXPECTED (mediaSource.sourceBuffers[0].buffered.end(0) > '1') OK

Set timestampOffset.
EXPECTED (segmentHelper.sourceBuffer.timestampOffset == '0') OK
RUN(segmentHelper.sourceBuffer.timestampOffset = 5)
EXPECTED (segmentHelper.sourceBuffer.timestampOffset == '5') OK

Append 1 second of data.
EXPECTED (mediaSource.sourceBuffers[0].buffered.length == '2') OK
EXPECTED (mediaSource.sourceBuffers[0].buffered.start(1) == '5') OK
EXPECTED (mediaSource.sourceBuffers[0].buffered.end(1) > '5') OK

Round-trip a precise timestampOffset
EXPECTED (segmentHelper.sourceBuffer.timestampOffset == '5') OK
RUN(segmentHelper.sourceBuffer.timestampOffset = 5.123456789)
EXPECTED (segmentHelper.sourceBuffer.timestampOffset == '5.123456789') OK

Remove SourceBuffer
RUN(mediaSource.removeSourceBuffer(segmentHelper.sourceBuffer))
EXPECTED (mediaSource.sourceBuffers.length == '0') OK
EXPECTED (mediaSource.sourceBuffers == '[object WebKitSourceBufferList]') OK

Test that append() throws an error after SourceBuffer has been removed.
Got expected exception InvalidStateError: Failed to execute 'append' on 'WebKitSourceBuffer': This SourceBuffer has been removed from its parent MediaSource.
Test that buffered throws an error after SourceBuffer has been removed.
Got expected exception InvalidStateError: Failed to read the 'buffered' property from 'WebKitSourceBuffer': This SourceBuffer has been removed from its parent MediaSource.
EXPECTED (mediaSource.sourceBuffers.length == '0') OK
Test that addSourceBuffer() throws and error when the MediaSource is not associated with a media element.
Got expected exception InvalidStateError: Failed to execute 'addSourceBuffer' on 'WebKitMediaSource': This MediaSource's readyState is not 'open'.
Test that setting timestampOffset throws an error when the MediaSource is not associated with a media element.
Got expected exception InvalidStateError: Failed to set the 'timestampOffset' property on 'WebKitSourceBuffer': This SourceBuffer has been removed from its parent MediaSource.
END OF TEST
